% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{uqq.rmse}
\alias{uqq.rmse}
\title{Compute RMSE of uniform QQ deviation}
\usage{
uqq.rmse(p, lwr = 0)
}
\arguments{
\item{p}{Numeric vector of p-values.}

\item{lwr}{Lower bound to truncate the p-value range (default = 0).}
}
\value{
A numeric scalar giving the RMSE.
}
\description{
Calculates the root mean squared error (RMSE) between observed p-values
and the expected uniform(0,1) distribution in a QQ plot.
}
\examples{
set.seed(1)
pvals = runif(100)
uqq.rmse(pvals)
uqq.rmse(pvals, lwr = 0.1)

}
